pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "paimonsdk"
version = "0.1.0"
description = "Python agent runtime kernel for custom agent loops."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.3",
    "python-dotenv>=1.0.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["paimonsdk*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
